bullion-tracker/coin_scraper/requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=4.5.0
